Feat/wlnp 12501 simplify asset price compose (#598)
* WLNP-12501: BSJ: Steams: Define new portfolio product type

improve ingest asset data; add price to asset model, remove duplication of price setup; add tests; fix PR recommendations;
Update CHANGELOG.md

---------

+++ stream-investment/investment-core/src/main/java/com/backbase/stream/investment/InvestmentDataValue.java
@@ -6,7 +6,7 @@ public interface InvestmentDataValue {
 
     long getTotalProcessedValues();
 
-    default <T> int getSize(List<T> list) {
+    default <T> long getSize(List<T> list) {
         return list != null ? list.size() : 0;
     }
